tests: smp_serial: rename test fragmented -> split packet
The term 'fragemented' describes the splitting on mcumgr layer. Here
it is meant that ONE mcumgr packet is split in the smp_serial transport
layer. The packet is in this sense not fragemented (which can happen
on top of the smp_serial transport splitting)

Add comments to make it more clear.